requirements.txt
streamlit
gspread
pandas
matplotlib
google-auth
google-auth-oauthlib